правка кода

+++ telegrammCategory/src/main/java/com/telegrammCategory/service/impl/CategoryServiceImpl.java
@@ -1,42 +1,45 @@
 package com.telegrammCategory.service.impl;
 
+
+
 import com.telegrammCategory.exception.ElemNotFound;
 import com.telegrammCategory.model.Category;
 import com.telegrammCategory.repository.CategoryRepository;
 import com.telegrammCategory.service.CategoryService;
-import lombok.RequiredArgsConstructor;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.Collection;
 
 
 /**  Сервис Категорий  */
-@RequiredArgsConstructor
+
+@AllArgsConstructor
 @Service
 public class CategoryServiceImpl implements CategoryService {
-    private  Category category;
     private CategoryRepository categoryRepository;
 
 
     @Override
     public String getCategoryLevel(int level) {
-        try {
-            Collection<String> getCategory = categoryRepository.findByParent(level);
-            return String.join("\n", getCategory);
-        }catch (ElemNotFound e){return "Это меню пустое";}
-
+        return String.join("\n", categoryRepository.findAllByParent(level));
     }
 
     @Override
-    public int getCategoryPreviousLevel(int level) {
-        return categoryRepository.findPreviousLevel(level);
+    public String getCategoryPreviousLevel(int level) {
+        return String.join("\n",categoryRepository.findPreviousLevel(level));
     }
 
     @Override
     public String greatCategory(int level, String name) {
-        int maxSeq = categoryRepository.findByParentAndMaxSeg(level);
+        int maxSeq =1;
+        try {
+            maxSeq = categoryRepository.findByParentAndMaxSeg(level).orElseThrow(ElemNotFound::new);
+        } catch (ElemNotFound e) {
+        }
+
         Category category1 = new Category();
-        category1.setSeq(category1.getSeq()+1);
+        category1.setParent(level);
+        category1.setSeq(maxSeq++);
         category1.setName(name);
         categoryRepository.save(category1);
         return getCategoryLevel(level);
@@ -47,19 +50,22 @@ public class CategoryServiceImpl implements CategoryService {
         Category category1 = new Category();
         category1.setSeq(1);
         category1.setName(name);
-        category1.setParent_node_id(id);
+        category1.setParent(id);
         return getCategoryLevel(id);
     }
 
     @Override
     public void deleteCategory(int id, int level) {
-        Category category1 = categoryRepository.findByParentAndSeg(level,id);
-        categoryRepository.delete(category1);
+        Category category1 = categoryRepository.findByParentAndSeg(level, id).orElseThrow(ElemNotFound::new);
+        if (category1 != null) {
+            categoryRepository.delete(category1);
+        }
+
     }
 
     @Override
     public int newLevel(Integer level, int value) {
-        Category category1 = categoryRepository.findByParentAndSeg(level,value);
+        Category category1 = categoryRepository.findByParentAndSeg(level,value).orElseThrow(ElemNotFound::new);
         category1.setName(category1.getName() + " >");
         categoryRepository.save(category1);
         return category1.getId();
